// generated/api-client/models/blog-detail-response-dto.ts
export interface BlogDetailResponseDto {
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'slug': string;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'subtitle': string | null;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'excerpt': string | null;
    /**
     * 
     * @type {number}
     * @memberof BlogDetailResponseDto
     */
    'readingTime': number | null;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'status': string;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'publishedAt': string | null;
    /**
     * 
     * @type {number}
     * @memberof BlogDetailResponseDto
     */
    'viewCount': number;
    /**
     * 
     * @type {number}
     * @memberof BlogDetailResponseDto
     */
    'likeCount': number;
    /**
     * 
     * @type {number}
     * @memberof BlogDetailResponseDto
     */
    'commentCount': number;
    /**
     * 
     * @type {number}
     * @memberof BlogDetailResponseDto
     */
    'bookmarkCount': number;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDetailResponseDto
     */
    'isPinned': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDetailResponseDto
     */
    'isFeatured': boolean;
    /**
     * 
     * @type {BlogSummaryResponseDtoAuthor}
     * @memberof BlogDetailResponseDto
     */
    'author': BlogSummaryResponseDtoAuthor;
    /**
     * 
     * @type {Array<MyBlogResponseDtoCoverImagesInner>}
     * @memberof BlogDetailResponseDto
     */
    'coverImages': Array<MyBlogResponseDtoCoverImagesInner>;
    /**
     * 
     * @type {Array<MyBlogResponseDtoTagsInnerTag>}
     * @memberof BlogDetailResponseDto
     */
    'tags': Array<MyBlogResponseDtoTagsInnerTag>;
    /**
     * 
     * @type {Array<MyBlogResponseDtoTagsInnerTag>}
     * @memberof BlogDetailResponseDto
     */
    'categories': Array<MyBlogResponseDtoTagsInnerTag>;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'updatedAt': string;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDetailResponseDto
     */
    'isLiked': boolean | null;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDetailResponseDto
     */
    'isBookmarked': boolean | null;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'content': string;
    /**
     * 
     * @type {number}
     * @memberof BlogDetailResponseDto
     */
    'wordCount': number | null;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'metaDescription': string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof BlogDetailResponseDto
     */
    'metaKeywords': Array<string>;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'canonicalUrl': string | null;
    /**
     * 
     * @type {number}
     * @memberof BlogDetailResponseDto
     */
    'shareCount': number;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'moderationStatus': string;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDetailResponseDto
     */
    'allowComments': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDetailResponseDto
     */
    'allowBookmarks': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDetailResponseDto
     */
    'allowLikes': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDetailResponseDto
     */
    'allowShares': boolean;
    /**
     * 
     * @type {string}
     * @memberof BlogDetailResponseDto
     */
    'lastEditedAt': string;
}