// generated/api-client/models/blog-list-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BlogListResponseDtoMeta } from './blog-list-response-dto-meta';
// May contain unused imports in some cases
// @ts-ignore
import { BlogSummaryResponseDto } from './blog-summary-response-dto';

/**
 * 
 * @export
 * @interface BlogListResponseDto
 */
export interface BlogListResponseDto {
    /**
     * 
     * @type {string}
     * @memberof BlogListResponseDto
     */
    'message': string;
    /**
     * 
     * @type {number}
     * @memberof BlogListResponseDto
     */
    'statusCode': number;
    /**
     * 
     * @type {Array<BlogSummaryResponseDto>}
     * @memberof BlogListResponseDto
     */
    'blogs': Array<BlogSummaryResponseDto>;
    /**
     * 
     * @type {BlogListResponseDtoMeta}
     * @memberof BlogListResponseDto
     */
    'meta': BlogListResponseDtoMeta;
}